Adding dataProvider
