+++ src/dataProvider.js
@@ -1,0 +1,22 @@
+/* eslint-disable no-await-in-loop */
+
+const token = '43aae801c5a8102a0468a7ce5398fac958611e6c';
+
+export async function getPrs(user, repo) {
+  let prsList = [];
+  let page = 1;
+  let promiseValue = [0];
+  while (promiseValue.length !== 0) {
+    const fetchResult = await fetch(`https://api.github.com/repos/${user}/${repo}/pulls?access_token=${token}&page=${page}&per_page=100&state=all&sort=created_at`);
+    const promise = fetchResult.json();
+    promiseValue = await promise.then(value => value);
+
+    prsList = prsList.concat(promiseValue);
+    page += 1;
+  }
+
+  return prsList;
+}
+
+
+export default { getPrs };
