+++ src/newContributorsPRs.js
@@ -1,25 +1,8 @@
 /* eslint-disable no-await-in-loop */
 
-// Use a Node.js core library
 import * as d3 from 'd3';
+import { getPrs } from './dataProvider';
 
-const token = '43aae801c5a8102a0468a7ce5398fac958611e6c';
-
-async function getPrs(user, repo) {
-  let prsList = [];
-  let page = 1;
-  let promiseValue = [0];
-  while (promiseValue.length !== 0) {
-    const fetchResult = await fetch(`https://api.github.com/repos/${user}/${repo}/pulls?access_token=${token}&page=${page}&per_page=100&state=all&sort=created_at`);
-    const promise = fetchResult.json();
-    promiseValue = await promise.then(value => value);
-
-    prsList = prsList.concat(promiseValue);
-    page += 1;
-  }
-
-  return prsList;
-}
 
 function numberPrsNewContributorsAcceptedDonutChart(
   notAcceptedPrs, acceptedPrs, size, divId,
